Consolidate gateway_invocation.py into gateway_adapter.py
Move all gateway invocation logic (HTTP calls, retry, context window
management, response parsing) from the separate gateway_invocation.py
into gateway_adapter.py, matching the pattern used by litellm_adapter.py.

invoke_via_gateway_and_handle_tools becomes
GatewayAdapter._run_tool_calling_loop (private method).
